Format & Lint fixes
