+++ src/components/LargeBookingTable.tsx
@@ -2,7 +2,15 @@ import React, { ChangeEvent } from 'react';
 import { BookingViewModel } from '../models/interfaces';
 import BookingTypeTag from '../components/utils/BookingTypeTag';
 import { TableDisplay, TableConfiguration } from '../components/TableDisplay';
-import { countNullorEmpty, getStatusColor, getStatusName, nameSortFn, notEmpty, onlyUnique, onlyUniqueById } from '../lib/utils';
+import {
+    countNullorEmpty,
+    getStatusColor,
+    getStatusName,
+    nameSortFn,
+    notEmpty,
+    onlyUnique,
+    onlyUniqueById,
+} from '../lib/utils';
 import { Typeahead } from 'react-bootstrap-typeahead';
 import { Col, Form } from 'react-bootstrap';
 import { Status } from '../models/enums/Status';
