+++ src/pages/api/calendar/index.ts
@@ -7,7 +7,7 @@ import { fetchSettings } from '../../../lib/db-access/setting';
 import { withSessionContext } from '../../../lib/sessionContext';
 import { getGlobalSetting } from '../../../lib/utils';
 import { CalendarResult } from '../../../models/misc/CalendarResult';
-import { fetchUserIdByNameTag } from '../../../lib/db-access/user';
+import { getNameTagsFromEventName, getUsersIdsFromEventName } from '../../../lib/calenderUtils';
 
 const calendarClient = calendar({
     version: 'v3',
@@ -86,24 +86,5 @@ const handler = withSessionContext(async (req: NextApiRequest, res: NextApiRespo
     }
     return;
 });
-
-const getNameTagsFromEventName = (name: string): string[] => {
-    // Get part of string within [] brackets
-    const match = name.match(/\[(.*?)\]/);
-    if (match) {
-        return match[1]
-            .split(',')
-            .map((x) => (x.includes(':') ? x.split(':')[1] : x))
-            .map((x) => x.trim());
-    }
-    return [];
-};
-
-const getUsersIdsFromEventName = async (name: string): Promise<number[]> => {
-    const nameTags = getNameTagsFromEventName(name ?? '');
-    const users = await Promise.all(nameTags.map(fetchUserIdByNameTag));
-
-    return users.filter((x) => x !== undefined).map((x) => x.id!);
-};
 
 export default handler;
