+++ src/pages/api/calendar/open-hours.ts
@@ -6,6 +6,7 @@ import { fetchSettings } from '../../../lib/db-access/setting';
 import { withSessionContext } from '../../../lib/sessionContext';
 import { getGlobalSetting } from '../../../lib/utils';
 import { CalendarResult } from '../../../models/misc/CalendarResult';
+import { getNameTagsFromEventName, getUsersIdsFromEventName } from '../../../lib/calenderUtils';
 
 const calendarClient = calendar({
     version: 'v3',
@@ -51,6 +52,8 @@ const mapCalendarResponse = (res: GaxiosResponse<calendar_v3.Schema$Events>): Pr
                 link: x.htmlLink ?? undefined,
                 start: x.start?.dateTime ?? x.start?.date ?? undefined,
                 end: x.end?.dateTime ?? x.start?.date ?? undefined,
+                initials: getNameTagsFromEventName(x.summary ?? ''),
+                workingUsersIds: await getUsersIdsFromEventName(x.summary ?? ''),
             })),
     );
 };
